chore: move util tests into a separate file
